fix migrations

+++ database/migrations/2022_02_27_204135_food_order.php
@@ -14,8 +14,8 @@ class FoodOrder extends Migration
     public function up()
     {
         Schema::create('food_order', function (Blueprint $table) {
-            $table->integer('order_id');
-            $table->integer('food_id');
+            $table->unsignedBigInteger('order_id');
+            $table->unsignedBigInteger('food_id');
             $table->integer('quantity');
             $table->foreign('order_id')->references('id')->on('order')->onDelete('cascade');
             $table->foreign('food_id')->references('id')->on('food')->onDelete('cascade');
